requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==1.1.1
werkzeug==2.3.7
schedule==1.2.0
qrcode[pil]==7.4.2
Pillow==10.0.0
requests==2.31.0
gunicorn==21.2.0